\name{sculpt3d-package}
\Rdversion{1.1}
\alias{sculpt3d-package}
\docType{package}
\title{
A simple toolbar GUI for brushing RGL plots
}
\description{
sculpt3d is a GTK+ toolbar that allows for more interactive control of a dataset inside the RGL plot window.  Controls for simple brushing, highlighting, labeling, and mouseMode changes are provided by point-and-click rather than through the R terminal interface.
}
\details{
\tabular{ll}{
Package: \tab sculpt3d\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2009-07-15\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Justin Donaldson

Maintainer: Justin Donaldson <jdonaldson@gmail.com>
}



\keyword{ package }
\seealso{

		\code{\link{sculpt3d}} : Create a toolbar to plot and interact with the given data \cr
		\code{\link{sculpt3d.selected}} : Retrieve a logical vector indicating currently selected points \cr
		\code{\link{sculpt3d.current}} : Retrieve a logical vector indicating currently visible points \cr
		\code{\link{sculpt3d.setCallback}} : set a callback function for selection/crop/deletion events \cr

}

