\name{sculpt3d.current}
\Rdversion{1.1}
\alias{sculpt3d.current}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Retrieve currently visible points from RGL plot
}
\description{
Returns a logical vector (of TRUE/FALSE) indicating if the given point index is currently displayed in the RGL interface.
}
\usage{
sculpt3d.current()
}
%- maybe also 'usage' for other objects documented here.

\value{
A vector of TRUE/FALSE values of equal length to the number of datapoints provided to sculpt3d().
}

\author{
Justin Donaldson
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{sculpt3d}}, 
\code{\link{sculpt3d.selected}} 
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ attribute }
